Fix el translation in popup
Correct calculation of el in el.popup. 'right' and 'down' directions appeared to miscalculate the x and y translations of the el respectively, which is corrected by using the offsets x2 and y2 in those cases instead of x and y.

This has only been tested in my original use case.